+++ Sources/StreamChat/Utils/MarkdownParser.swift
@@ -33,23 +33,24 @@ public enum MarkdownParser {
     /// - Parameters:
     ///   - markdown: The string that contains the Markdown formatting.
     ///   - attributes: The attributes to use for the whole string.
+    ///   - options: Options that affect how the Markdown string is parsed and styled.
     ///   - inlinePresentationIntentAttributes: The closure for customising attributes for inline presentation intents.
     ///   - presentationIntentAttributes: The closure for customising attributes for presentation intents. Called for quote, code, list item, and headers.
     public static func style(
         markdown: String,
+        options: ParsingOptions,
         attributes: AttributeContainer,
         inlinePresentationIntentAttributes: (InlinePresentationIntent) -> AttributeContainer?,
         presentationIntentAttributes: (PresentationIntent.Kind, PresentationIntent) -> AttributeContainer?
     ) throws -> AttributedString {
-        let options = AttributedString.MarkdownParsingOptions(
-            allowsExtendedAttributes: true,
-            interpretedSyntax: .full,
-            failurePolicy: .returnPartiallyParsedIfPossible,
-            languageCode: nil
-        )
         var attributedString = try AttributedString(
             markdown: markdown,
-            options: options
+            options: AttributedString.MarkdownParsingOptions(
+                allowsExtendedAttributes: true,
+                interpretedSyntax: .full,
+                failurePolicy: .returnPartiallyParsedIfPossible,
+                languageCode: nil
+            )
         )
         
         attributedString.mergeAttributes(attributes)
@@ -121,7 +122,7 @@ public enum MarkdownParser {
                 if blockStyling.isOrdered == true {
                     blockStyling.prependedString.append("\(listItemOrdinal).\t")
                 } else {
-                    blockStyling.prependedString.append("・\t")
+                    blockStyling.prependedString.append("\u{2022}\t")
                 }
                 // Extra space when list's last item
                 if let previousBlockStyling, previousBlockStyling.listId != blockStyling.listId {
@@ -132,7 +133,7 @@ public enum MarkdownParser {
             if blockStyling.succeedingNewlineCount > 0, let previousBlockStyling {
                 let newlineString = String(repeating: "\n", count: blockStyling.succeedingNewlineCount)
                 let insertedString = AttributedString(newlineString, attributes: attributes)
-                attributedString.insert(insertedString, at: previousBlockStyling.range.lowerBound)
+                attributedString.insertSafely(insertedString, at: previousBlockStyling.range.lowerBound)
             }
             // Additional attributes
             if let attributes = blockStyling.mergedAttributes {
@@ -141,20 +142,47 @@ public enum MarkdownParser {
             // Inserting additional characters (list items etc)
             if !blockStyling.prependedString.isEmpty {
                 let attributes = attributes.merging(blockStyling.mergedAttributes ?? AttributeContainer())
-                let insertedString = AttributedString(blockStyling.prependedString, attributes: attributes)
-                attributedString.insert(insertedString, at: range.lowerBound)
+                if options.layoutDirectionLeftToRight {
+                    let insertedString = AttributedString(blockStyling.prependedString, attributes: attributes)
+                    attributedString.insertSafely(insertedString, at: range.lowerBound)
+                } else {
+                    let insertedString = AttributedString(blockStyling.prependedString.reversed(), attributes: attributes)
+                    attributedString.insertSafely(insertedString, at: range.upperBound)
+                }
             }
             // Spacing before the block
             if blockStyling.precedingNewlineCount > 0, attributedString.startIndex != range.lowerBound {
                 let newlineString = String(repeating: "\n", count: blockStyling.precedingNewlineCount)
                 let insertedString = AttributedString(newlineString, attributes: attributes)
-                attributedString.insert(insertedString, at: range.lowerBound)
+                attributedString.insertSafely(insertedString, at: range.lowerBound)
             }
             
             previousBlockStyling = blockStyling
         }
                 
         return attributedString
+    }
+}
+
+@available(iOS 15, *)
+extension MarkdownParser {
+    /// Options that affect how the Markdown string is parsed and styled.
+    public struct ParsingOptions {
+        public init(layoutDirectionLeftToRight: Bool = true) {
+            self.layoutDirectionLeftToRight = layoutDirectionLeftToRight
+        }
+        
+        /// Affects insertion index for additional characters like bullets and numbers for lists.
+        public var layoutDirectionLeftToRight = true
+    }
+}
+
+@available(iOS 15, *)
+private extension AttributedString {
+    mutating func insertSafely(_ s: some AttributedStringProtocol, at index: AttributedString.Index) {
+        // Inserting at the end index is same as appending
+        guard index >= startIndex, index <= endIndex else { return }
+        insert(s, at: index)
     }
 }