+++ Sources/StreamChatUI/Appearance+Formatters/MarkdownFormatter.swift
@@ -43,10 +43,11 @@ open class DefaultMarkdownFormatter: MarkdownFormatter {
     }
 
     open func format(_ string: String) -> NSAttributedString {
-        if #available(iOS 15, *) {
+        if #available(iOS 15, *), !string.isEmpty {
             do {
                 let attributedString = try MarkdownParser.style(
                     markdown: string,
+                    options: .init(layoutDirectionLeftToRight: UITraitCollection.current.layoutDirection == .leftToRight),
                     attributes: AttributeContainer(defaultAttributes),
                     inlinePresentationIntentAttributes: inlinePresentationIntentAttributes(for:),
                     presentationIntentAttributes: presentationIntentAttributes(for:in:)
@@ -65,10 +66,11 @@ open class DefaultMarkdownFormatter: MarkdownFormatter {
     // MARK: - Styling Attributes
     
     private var colorPalette: Appearance.ColorPalette { Appearance.default.colorPalette }
+    private var fonts: Appearance.Fonts { Appearance.default.fonts }
     
     private var defaultAttributes: [NSAttributedString.Key: Any] {
         [
-            .font: UIFont.font(forMarkdownFont: styles.bodyFont),
+            .font: UIFont.font(forMarkdownFont: styles.bodyFont, defaultFont: fonts.body),
             .foregroundColor: styles.bodyFont.color ?? Appearance.default.colorPalette.text
         ]
     }
@@ -78,8 +80,7 @@ open class DefaultMarkdownFormatter: MarkdownFormatter {
         switch inlinePresentationIntent {
         case .code:
             let attributes: [NSAttributedString.Key: Any] = [
-                // Inline currently does not have background color, although many editors prefer to do this
-                .font: UIFont.font(forMarkdownFont: styles.codeFont, monospaced: true),
+                .font: UIFont.font(forMarkdownFont: styles.codeFont, defaultFont: fonts.body, monospaced: true),
                 .foregroundColor: styles.codeFont.color
             ].compactMapValues { $0 }
             return AttributeContainer(attributes)
@@ -90,7 +91,7 @@ open class DefaultMarkdownFormatter: MarkdownFormatter {
     }
     
     @available(iOS 15, *)
-    private func presentationIntentAttributes(for presentationKind: PresentationIntent.Kind, in presentationIntent: PresentationIntent) -> AttributeContainer {
+    private func presentationIntentAttributes(for presentationKind: PresentationIntent.Kind, in presentationIntent: PresentationIntent) -> AttributeContainer? {
         switch presentationKind {
         case .blockQuote:
             return AttributeContainer([
@@ -98,8 +99,7 @@ open class DefaultMarkdownFormatter: MarkdownFormatter {
             ])
         case .codeBlock:
             let attributes: [NSAttributedString.Key: Any] = [
-                .backgroundColor: colorPalette.background2,
-                .font: UIFont.font(forMarkdownFont: styles.codeFont, monospaced: true),
+                .font: UIFont.font(forMarkdownFont: styles.codeFont, defaultFont: fonts.body, monospaced: true),
                 .foregroundColor: styles.codeFont.color
             ].compactMapValues { $0 }
             return AttributeContainer(attributes)
@@ -108,22 +108,22 @@ open class DefaultMarkdownFormatter: MarkdownFormatter {
             let foregroundColor: UIColor?
             switch level {
             case 1:
-                font = UIFont.font(forMarkdownFont: styles.h1Font, textStyle: .title1)
+                font = UIFont.font(forMarkdownFont: styles.h1Font, defaultFont: fonts.title, textStyle: .title1)
                 foregroundColor = styles.h1Font.color
             case 2:
-                font = UIFont.font(forMarkdownFont: styles.h2Font, textStyle: .title2)
+                font = UIFont.font(forMarkdownFont: styles.h2Font, defaultFont: fonts.title2, textStyle: .title2)
                 foregroundColor = styles.h2Font.color
             case 3:
-                font = UIFont.font(forMarkdownFont: styles.h3Font, textStyle: .title3)
+                font = UIFont.font(forMarkdownFont: styles.h3Font, defaultFont: fonts.title3, textStyle: .title3)
                 foregroundColor = styles.h3Font.color
             case 4:
-                font = UIFont.font(forMarkdownFont: styles.h4Font, textStyle: .headline)
+                font = UIFont.font(forMarkdownFont: styles.h4Font, defaultFont: fonts.headline, textStyle: .headline)
                 foregroundColor = styles.h4Font.color
             case 5:
-                font = UIFont.font(forMarkdownFont: styles.h5Font, textStyle: .subheadline)
+                font = UIFont.font(forMarkdownFont: styles.h5Font, defaultFont: fonts.subheadline, textStyle: .subheadline)
                 foregroundColor = styles.h5Font.color
             default:
-                font = UIFont.font(forMarkdownFont: styles.h6Font, textStyle: .footnote)
+                font = UIFont.font(forMarkdownFont: styles.h6Font, defaultFont: fonts.footnote, textStyle: .footnote)
                 foregroundColor = styles.h6Font.color ?? colorPalette.subtitleText
             }
             if let foregroundColor {
@@ -136,7 +136,7 @@ open class DefaultMarkdownFormatter: MarkdownFormatter {
                 .paragraphStyle: listItemParagraphStyle(forIndentationLevel: presentationIntent.indentationLevel)
             ])
         default:
-            return AttributeContainer()
+            return nil
         }
     }
     
@@ -153,11 +153,15 @@ open class DefaultMarkdownFormatter: MarkdownFormatter {
 private extension UIFont {
     static func font(
         forMarkdownFont markdownFont: MarkdownFont,
+        defaultFont: UIFont,
         textStyle: TextStyle = .body,
         monospaced: Bool = false
     ) -> UIFont {
+        if !markdownFont.hasFontChanges && !monospaced {
+            return UIFontMetrics(forTextStyle: textStyle).scaledFont(for: defaultFont)
+        }
         // Default
-        var descriptor = UIFontDescriptor.preferredFontDescriptor(withTextStyle: textStyle)
+        var descriptor = defaultFont.fontDescriptor
         if monospaced, let updatedDescriptor = descriptor.withDesign(.monospaced) {
             descriptor = updatedDescriptor
         }